Refactor subset selection
